[Silver IV] Title: 사이클 단어, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/1544. 사이클 단어/사이클 단어.cc
@@ -34,7 +34,7 @@ void	find_string(string s)
 	{
 		if (word[i].length() == str_size)
 		{
-			size_t idx = s.find(word[i][0]);
+			int idx = s.find(word[i][0]);
 			while (idx != string::npos)
 			{
 				if (s.substr(idx) == word[i].substr(0, str_size - idx))
